Refactor way of passing data to pdf
