+++ src/components/DayOverview/DayOverview.tsx
@@ -16,6 +16,12 @@ export type ParcelOfSpecificDateAndLocation = {
     > | null;
 };
 
+export type DayOverviewData = {
+    date: Date;
+    location: string;
+    data: ParcelOfSpecificDateAndLocation[];
+};
+
 export const getCurrentDate = (date: Date, hyphen: boolean = false): string => {
     const year = date.getFullYear();
     const month = (date.getMonth() + 1).toString().padStart(2, "0");
@@ -50,17 +56,44 @@ const getParcelsOfSpecificDateAndLocation = async (
     return data;
 };
 
+const collectionCentreToAbbreviation = (
+    collectionCentre: Schema["parcels"]["collection_centre"]
+): string => {
+    switch (collectionCentre) {
+        case "Brixton Hill - Methodist Church":
+            return "BH_MC";
+        case "Clapham - St Stephens Church":
+            return "CLP_SC";
+        case "N&B - Emmanuel Church":
+            return "NAB_EC";
+        case "Streatham - Immanuel & St Andrew":
+            return "STM_IS";
+        case "Vauxhall Hope Church":
+            return "VHC";
+        case "Waterloo - Oasis":
+            return "WAT_OA";
+        case "Waterloo - St George the Martyr":
+            return "WAT_SG";
+        case "Waterloo - St Johns":
+            return "WAT_SJ";
+        case "Delivery":
+            return "Delivery";
+        default:
+            throw new Error("Invalid location");
+    }
+};
+
 const DayOverview = async ({ text, date, location }: Props): Promise<React.ReactElement> => {
     const parcelsOfSpecificDate = await getParcelsOfSpecificDateAndLocation(date, location);
+    const dateString = getCurrentDate(date);
+    const fileName = `DayOverview_${dateString}_${collectionCentreToAbbreviation(location)}.pdf`;
+    const data = {
+        date: date,
+        location: location,
+        data: parcelsOfSpecificDate,
+    };
 
-    return (
-        <DayOverviewButton
-            date={date}
-            location={location}
-            data={parcelsOfSpecificDate}
-            text={text}
-        />
-    );
+    return <DayOverviewButton data={data} text={text} fileName={fileName} />;
 };
 
 export default DayOverview;
